requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
Pillow>=10.0
opencv-python>=4.8
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
